# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Sangli Soybean Yield & Market Advisory API","description":"Production-ready FastAPI backend providing harvest prediction, yield outlook, and market storage timing for Sangli farmers.","version":"0.2.0"},"paths":{"/api/health":{"get":{"tags":["System"],"summary":"Get Health","description":"Returns application health, model version, training timestamp, and market data as-of date.","operationId":"get_health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/options":{"get":{"tags":["Options"],"summary":"Get Options","description":"Returns valid dropdown options: talukas, crops, soils, varieties.","operationId":"get_options_api_options_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Options Api Options Get"}}}}}}},"/api/model-card":{"get":{"tags":["Model Card"],"summary":"Get Model Card","description":"Returns model metadata, evaluation results, and assumptions (JSON safe).","operationId":"get_model_card_api_model_card_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Model Card Api Model Card Get"}}}}}}},"/api/market":{"get":{"tags":["Market"],"summary":"Get Market Report","description":"Returns market price analysis and 6-month storage vs immediate selling plan.","operationId":"get_market_report_api_market_get","parameters":[{"name":"storage","in":"query","required":false,"schema":{"type":"number","minimum":0.0,"description":"Monthly storage cost in INR per quintal","default":15.0,"title":"Storage"},"description":"Monthly storage cost in INR per quintal"},{"name":"interest","in":"query","required":false,"schema":{"type":"number","maximum":0.1,"minimum":0.0,"description":"Monthly interest rate (e.g. 0.01 for 1%)","default":0.01,"title":"Interest"},"description":"Monthly interest rate (e.g. 0.01 for 1%)"},{"name":"msp","in":"query","required":false,"schema":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"description":"Optional Minimum Support Price in INR per quintal","title":"Msp"},"description":"Optional Minimum Support Price in INR per quintal"},{"name":"today","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Optional date override for analysis","title":"Today"},"description":"Optional date override for analysis"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Market Report Api Market Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/advisory":{"post":{"tags":["Advisory"],"summary":"Create Advisory","description":"Generates tailored harvest, yield outlook, and market storage advisory for a soybean farmer.\nReturns JSON matching the model contract.","operationId":"create_advisory_api_advisory_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvisoryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Create Advisory Api Advisory Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/harvest-feedback":{"post":{"tags":["Harvest Feedback & Retraining"],"summary":"Submit Harvest Feedback","description":"Records real-world farmer harvest observations to calibrate the harvest date model.\nAppends observation to data/harvest_feedback.csv.","operationId":"submit_harvest_feedback_api_harvest_feedback_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HarvestFeedbackRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HarvestFeedbackResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/retrain-harvest":{"post":{"tags":["Harvest Feedback & Retraining"],"summary":"Retrain Harvest Model","description":"Retrains the harvest date calibration using logged farmer feedback.\nProtected by the X-API-Key header.\nRequires at least 3 valid observations.\nDeep-copies the model before mutation, then atomically saves and hot-swaps.","operationId":"retrain_harvest_model_api_retrain_harvest_post","parameters":[{"name":"X-API-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetrainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdvisoryRequest":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name","description":"Farmer's name","default":""},"taluka":{"type":"string","title":"Taluka","description":"Taluka in Sangli district"},"crop":{"type":"string","title":"Crop","description":"Crop name (Soybean)","default":"Soybean"},"sowing_date":{"type":"string","format":"date","title":"Sowing Date","description":"Sowing date (YYYY-MM-DD)"},"soil":{"type":"string","title":"Soil","description":"Soil type from options"},"variety":{"type":"string","title":"Variety","description":"Soybean variety (early, medium, late or full description)"},"acres":{"type":"number","exclusiveMinimum":0.0,"title":"Acres","description":"Farm size in acres (must be > 0)"},"storage_cost":{"type":"number","minimum":0.0,"title":"Storage Cost","description":"Monthly storage cost in INR per quintal (>= 0)","default":15.0},"interest_rate":{"type":"number","maximum":0.1,"minimum":0.0,"title":"Interest Rate","description":"Monthly interest rate (between 0 and 0.1)","default":0.01},"today":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Today","description":"Optional current date override for demo/testing"}},"type":"object","required":["taluka","sowing_date","soil","variety","acres"],"title":"AdvisoryRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HarvestFeedbackRequest":{"properties":{"sowing_date":{"type":"string","format":"date","title":"Sowing Date","description":"Sowing date (YYYY-MM-DD)"},"harvest_date":{"type":"string","format":"date","title":"Harvest Date","description":"Harvest date (YYYY-MM-DD)"},"variety":{"type":"string","title":"Variety","description":"Variety: early, medium, or late"},"taluka":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Taluka","description":"Optional Sangli taluka"}},"type":"object","required":["sowing_date","harvest_date","variety"],"title":"HarvestFeedbackRequest"},"HarvestFeedbackResponse":{"properties":{"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"},"feedback":{"additionalProperties":true,"type":"object","title":"Feedback"}},"type":"object","required":["status","message","feedback"],"title":"HarvestFeedbackResponse"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"model_version":{"type":"string","title":"Model Version"},"trained_at":{"type":"string","title":"Trained At"},"market_data_as_of":{"type":"string","title":"Market Data As Of"}},"type":"object","required":["status","model_version","trained_at","market_data_as_of"],"title":"HealthResponse"},"RetrainResponse":{"properties":{"status":{"type":"string","title":"Status"},"calibrated":{"type":"boolean","title":"Calibrated"},"n_obs":{"type":"integer","title":"N Obs"},"message":{"type":"string","title":"Message"}},"type":"object","required":["status","calibrated","n_obs","message"],"title":"RetrainResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}